render helper: falsy contexts no longer treated as absent
    {{render 'something' falsyValue1}} 
    {{render 'something' falsyValue2}}

The above used to throw an error, treating the falsy values
as invocations of the `render` helper without 
a context, which you're not allowed to do if
rendering more than once. Now the above will not
throw, and the falsy context will be passed to
the generated controller as the model.