Smarter linkTo observers
Minor refactor to clean up some linkTo code and fix an inconsistency
whereby `{{linkTo 'elsewhere' this}}` would not automatically update
when the present route's ObjectController's content would change,
since the `this` was only watching the ObjectController, and not its
proxied content object. There was already code that special cased
the resolution of ObjectControllers' content, but it was missing from
the code that sets up the observers; now the same code is shared between
both code paths. Hooray.

Also removed the short-lived and probably unnecessary flag
HELPER_PARAM_LOOKUPS, which was only used for the linkTo,
and enabled linkTo parameters, routeName included, to be
bound paths when unquoted.

Thanks to @abahdanovich for pointing this out.