accessors: prevent duplicate warnings on get/set with dotted keys
`accessors.js` defines `get` and `set` and then, if `USE_ACCESSORS`
is falsy, it redefines them in terms of the original defenitions.
Whenever the second versions fell through to the original, they would
emit duplicate deprecation warnings about dotted keys.

This commit moves the deprecation warnings in the second versions of
`get` and `set` so that they only occur when *not* falling through to
the original definitions.