Don't force a layout on registered components
Currently, components, even when registered via Ember.Handlebars.helper,
are forced to have a corresponding layout name.

This makes it impossible to have a layout-less component (wrapping a
simple tag with some event handlers, like <input>). In particular, this
make it very awkward to subclass Ember.TextField.

This change preserves the current behavior if a component is built from
a template named components/*, but doesn't force a template on
components registered manually using Ember.Handlebars.helper.