Ember.computed.sort compares proxy correctly.
When an item property changes the item is removed and re-added, thus resorted.
When being removed, it is necessary to binary search using the previous value,
as that is what's determined the ordering.  However, because we coalesce changes
(and thus move them to a subsequent run loop) items will have their *new*
values.  Because of this, `Ember.computed.sort` uses a proxy object in this
case.

This commit makes the proxy object correctly compare itself to the proxied
object as 0 (ie equal).

This addresses issue 1 in #3273 but does not fully address issue 2.