`Ember.computed.sort` fallback to guids.
It is now legal for a user's sorting to result in different items counted as
equal for purposes of sorting.  Obviously the relative order of items that are
sort-equal is not guaranteed.

The implementation simply uses guids as a deterministic tie-breaker.

This addresses one of the bugs demonstrated by issue 2 in #3273.