[BUGFIX beta] Fixes issue when content is undefined for Ember.Select with optgroup
If the content given is undefined at the time of initialization, and
later changed,
`Ember.Select` correctly renders an empty select and then updates the
content  [Example](http://jsbin.com/epodAwU/14/edit),
But `Ember.Select` with `optgroup` breaks.
[Example](http://jsbin.com/IJaToVO/1/edit).

This breakage may happen, when the content for the select is fetched
via an ajax call to the server.

This PR fixes it.