[FEATURE ember-testing-wait-hooks]
In integration testing, sometimes it's necessary to make `wait` respect
other kinds of pending actions that are not in its default
implementation. For example, if your application is waiting for a CSS3
transition or an IndexDB action, the default `wait` implementaiton will
ignore those things and move on with the test prematurely.

This change lets you register additional functions that `wait` will call
to see whether it's really time to move on.