[BUGFIX] Fix an issue with concatenatedProperties.
This commit fixes an issue with how contactedProperties are applied that
can result in subclasses sharing a reference to the same array, in a
surprising way. As an example, the failing test provided here shows
how two Ember.App subclasses shares were sharing inadvertently
sharing initializers.

Thanks to @kselden and @stefanpenner for code review.