[BUGFIX beta] Fix edge case in `TrackedArray`.
Fixes http://jsbin.com/EFezAP/2

Fix an edge case in delete composition for a delete that right-trims an insert.
This results in a delete followed by an insert that should both compose to a
no-op, but the insert was being erroneously left behind.

Also ignore no-op adds and deletes (ie add 0 items, delete 0 items).