Use consistent versioning.
Update to latest `ember-dev` and utilize a consistent version number for
all builds. This means that the version comment and the `Ember.VERSION`
constant will be the same.

The `/VERSION` file in each branches directory will indicate the next
tagged release version along with custom metadata (`+pre` for release/beta and
`+canary` for master) indicating it is a prerelease version.

See
https://github.com/emberjs/ember-dev/commit/4018355f5ee9c2dd0533e78c9cfa15dd72497096
for a more complete explanation.