[BUGFIX beta] Increase precedence of route.controllerName
Currently a controller specified for a route via the `controllerName`
property is considered too late when the controller for the route is
determined: when a controller with the routeName already exists - for
example it has been used elsewhere in the app before - then it is used,
regardless of `controllerName` being present on the route.

This fix moves the check for a specified `controllerName` before the
route-name-controller-lookup.