[BUGFIX beta] {{partial}} helper now works with bound params
{{partial name}} will look up `name` on
the present template context and render the partial
whose name is the present value of `name`. If `name`
changes, the partial gets re-rendered. If name is
falsy, nothing gets rendered. If name refers to a
non-existent template/partial, the same assertion you'd
see if you'd explicitly passed in an invalid template/partial
name in string form will fire.