FEATURE: Application Can Ignore Specified Events
Rob Harper pointed out in his talk on performance with Ember on mobile
devices, Ember listens to a large number of events, some of which we may
not be interested in.

This pull request introduces a property on Ember.Application that allows
us to ignore certain events without having to fiddle with, and reopen
the Ember.EventDispatcher class.

References:

* http://www.robharper.ca/ember-on-mobile/performance.html