[DOC beta] Assert if template for {{render}} not found.
Added an assertion if the template name provided to `{{render}}` was not
found.

Two tests were previously not working properly, and have been fixed. They registered
templates as `"blog/post"`, which is never looked up with `{{render}}` (as any
slashes in the name are translated to a period before this lookup
happens).

Resolves #3926.