[FEATURE ember-routing-bound-action-name]
The `{{action}}` helper will now use a non-quoted parameter and perform
a bound property lookup against the action's target at the time the
event is triggered.

If a non-quoted param is used, but the property lookup returns
`undefined` a helpful assertion is made, and the specified parameter is
used as a plain string action name (thus preventing this from breaking
anyones applications).

Also, cleans up a few unrelated non-quoted uses of `{{action}}` in the
test suite.