[BUGFIX beta] Allow Component to lookup a templateName if specified.
Prior to this change, specifying `templateName` to a component would not
cause that template to be used (since we were overriding the
`Ember.View.template` function with `null`).

This change uses the logic from `Ember.View.template`, which is not
ideal, but I think that these custom docs are pretty important to have.
(They help instruct users that the block passed to a component is
interpreted as its template.)