[FEATURE query-params-new] alias computed.reads as computed.oneWay
oneWay is ambiguous. Which way?

Ember.computed.alias is a way more semantically meaningful alias of
Ember.computed.oneWay.

Example: http://jsbin.com/ucanam/3000

Putting this in query params because I'd like to use this
in some of the examples for query params; seems like it'll be
common lightweight pattern for buffering small changes, and
oneWay is just enough of a mental hurdle to necessitate this.