Allow reuse of Application.buildContainer without full Application.
This will allow usage of `Ember.Application.buildContainer` without
needing providing a namespace (or an Ember.Application instance).

The primary impetus of this is to prevent from having to create one-off
container look-alikes for unit testing. This also allows us to
consolidate the knowledge of what things are registered and setup by
default to a single place.