[FEATURE ember-views-bindable-attributes]
Any properties listed in `bindableAttributes` that exist on the
at the time of the first render will be bound (and observed).
However, any properties that do not exist will not have observers
setup.

The main difference here from `attributeBindings` is the fact that the
observers are only setup when needed.

The `Ember.BindableAttributesMixin` is mixed into `Ember.TextSupport` by
default, and all HTML5 attributes for `<input type='text'>` and `<textarea>`.