[query-params-new] Fix issue where active class was not being applied to parent route
This updates ember with the folliwing PR in router.js

https://github.com/tildeio/router.js/pull/78

The basic gist here is to only check for changes in query paramters that are passed in
to the `isActive` method. Additional query paramters that are active in the current state are
discared. This is because we want parent routes to be active but they were not because the
removal of query parsamters between the two routes caused the `isActive` method to return
`false` in router.js.