[BUGFIX beta] Use better ajax events for counters
Using `ajaxSend` and `ajaxComplete` instead of `ajaxStart` and
`ajaxStop`. See http://api.jquery.com/ajaxSend/ and
http://api.jquery.com/ajaxComplete/ for more information on each one.

It turns out that `ajaxStart` and `ajaxStop` are only called when the
request in question is the only one running. This is problematic for the
`wait()` helper. Changing to events that are called on _every_ request
fixes this problem.