[BUGFIX beta] The `each` helper checks that the metamorph tags have the same parent.
In some cases, the browser may add or fix tags, which change the
parentage of metamorph tags (This problem happens frequently when the
developer does not include the TBODY tag inside a table for example).
This change prevents the framework to update or clean the underlying DOM
elements.

This helps to detect the issue described in #4273.

Adds `Ember.runInDebug` to call a function in debug mode only