fix action bubbling for components with a component as parent
Fixes an issue where nesting a component inside another component would
cause actions to be swallowed and not delegated up to the route like
they should be. Ember.TargetActionSupport defines `targetObject` as
a computed property. This commit allows action bubbling to fall back
to that property instead of stopping the bubbling chain at the parent
component.

fixes GH-4430