Don't raise error when a destroyed array is assigned to ArrayProxy
Sometimes due to timing in test suites, an array that has become
destroyed (most likely due to a call to Store#unloadAll in Ember Data)
is assigned to an array proxy. This leads to an intermittent problem
where an assertion is raised due to a destroyed array-like object
no longer being considered an array by Ember.

The root cause is that Ember.isArray relies on Ember.Array.detect to
determine "Array-ness", and when Ember.Array.detect is used on a
destroyed object false is returned because the meta information has
been removed.