Enable ember-debug to be loaded for non-prod builds.
The issue is that ember-debug has moved to an ES6 package and even
though it exists in the build output, it is never required so the global
`Ember` object is never extended.

This is a short-term fix, but the longer term solution
is to remove the `ember-debug` package altogether (in favor of a better
build process).