Ensure that Ember.DataAdapter exists after Ember is loaded.
This fixes the underlying issue that #4505 was attempting to resolve
(but failed). #4505 ensured that the globals were setup properly upon
creation of an Ember.Application, but it unfortunately doesn't fix
Ember Data's initial boot.

After this PR the globals are all exported as part of the Ember boot
process (instead of during application creation time).