[BUGFIX beta] Only set pendingAjaxRequests once.
This was introduced in https://github.com/emberjs/ember.js/pull/4435,
where we are setting `pendingAjaxRequests` everytime
Ember.setupForTesting is called. This causes assertions in an EAK style
test suite where a new application is created for each test, since their
might be a pending ajax request that was left over from a prior test.