[FEATURE ember-routing-nested-resource-can-inherit-namespace]
Specifying `resetNamespace: false` in a nested resource declaration will automatically prefix the
new route with the parent's name (similar to `route`). This is much closer in line with the way
many people assume it works out of the box, and is likely a better default. The actual default can
not be changed until Ember 2.0 is release due to SemVer compatibility issues.

Example:

```javascript
var router = Router.map(function(){
  this.resource('bleep', function(){
    this.resource('bloop', {resetNamespace: false});
  });
});

// will create `bleep.bloop` route (without resetNamespace it would create `bloop` route)
```