[BUGFIX beta] Ensure context is unchanged when using keywords with itemController.
Prior to this change the following code would bind `this` in the
template block to the itemController's content (even though the keyword
form of `{{each}}` is used).

This change sets the child view's `_context` property to the current context when using
the keyword form of `{{each}}`, and a couple of confirming tests to
demonstrate using `itemController` specified in either the
`ArrayController` or the template directly.