The example of extending Ember.TextField with required is not relevant anymore
Since the `required` attribute is part of the recognised attributes as listed on line 27, the example about extending Ember.TextField shouldn't be about extending with `required`.

However I am under the impression that default "coming soon" behaviour is to accept all attributes by default instead of specifying them. If that's the case the whole Extension paragraph might be irrelevant.