[BUGFIX beta] reduceComputed detect retain:n better.
Fix #4620.

reduceComputed tries to skip walking all observer contexts when an item property
changes if the current tracked array state is [retain:n].  However, this check
occurs while applying the operations.  If [delete:m retain:(n-m)] is applied,
when the retain operation is visisted it will have been updated to retain:n.

This commit updates the check so that retain:n is only checked for the *first*
operation.

Conflicts:
	packages_es6/ember-runtime/tests/computed/reduce_computed_test.js