[DOC beta] Add assertion for lowerCamelCase model name
The naming convention as pointed out
[here](http://emberjs.com/guides/models/defining-models/) is to use the
`lowerCamelCase` version of the model name, though `store.find('MyModel', 1)`
works, since the name is `classify`-ed in the
[`Ember.Resolver#resolveModel`](http://git.io/1HDwJQ) method. This is
problematic since the passed model name is set as `typeKey` on the
factory in the [`Store#modelFor`](http://git.io/eb3iPQ) method, which in
turn leads to bugs, like reported in https://github.com/emberjs/data/issues/1890.

This PR adds an assertion message to the `Ember.Resolver#resolveModel`
hook, so the misuse of the naming convention is pointed out.